pyproject.toml
[tool.poetry]
name = "nonebot-plugin-revgptapi"
version = "0.1.0"
description = "通过GPTApi在QQ上与ChatGPT对话，支持GPT4"
authors = ["铂屑 <80164656+boxie123@users.noreply.github.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "nonebot_plugin_revgptapi"}]
homepage = "https://github.com/boxie123/nonebot-plugin-revGPTApi"
repository = "https://github.com/boxie123/nonebot-plugin-revGPTApi"

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = {extras = ["fastapi"], version = "^2.0.0rc4"}
nonebot-adapter-onebot = "^2.2.3"
aiofiles = "^23.1.0"
revchatgpt = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
